Chore: Move file to right place
